Upgrade to version 0.3-beta
Upgrade function to read previous IP address. Reading is done from the local DNS server. This update eliminates the problem of crashing the program when the retrieval function of the current address does not receive a response, ie when it loads an empty wan_ip file. And requires an additional dns_ip file